requirements2.txt
celery
email-reply-parser
akismet
markdown
beautifulsoup4
lxml
simplejson
pytz
six
pinax_teams
djangorestframework
django-model-utils
django-encrypted-model-fields